frete_service/requirements.txt
Django>=5.1
djangorestframework
psycopg2-binary
pydantic==2.9.2